+++ app/controllers/settings/users_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Settings::UsersController < ApplicationController
-  before_action :authenticate_self_hosted!
+  before_action :authenticate_self_hosted!, only: [:export, :import]
   before_action :authenticate_admin!, except: [:export, :import]
   before_action :authenticate_user!, only: [:export, :import]